+++ Game/Scripts/PlayerCameraRotationVerticalScript.cpp
