+++ Game/Scripts/PlayerCameraRotationVerticalScript.h
