+++ Source/Modules/ModuleInput.cpp
@@ -135,7 +135,10 @@ update_status ModuleInput::Update()
             break;
 
         case SDL_KEYDOWN:
-            keysState[sdlEvent.key.keysym.scancode] = KeyState::DOWN;
+            if (keysState[sdlEvent.key.keysym.scancode] != KeyState::REPEAT)
+            {
+                keysState[sdlEvent.key.keysym.scancode] = KeyState::DOWN;
+            }
             break;
         case SDL_KEYUP:
             keysState[sdlEvent.key.keysym.scancode] = KeyState::UP;
